Fixed errors
`Car` entity fields are now private
Fixed OneToOne relationship between `Car` and `User` entities
Changes have been made to `HQL queries` to prevent the problem of creating N+1 queries instead of one

+++ 2.2.1. spring_hibernate/src/main/java/hiber/service/UserServiceImp.java
@@ -29,7 +29,7 @@ public class UserServiceImp implements UserService {
 
     @Transactional(readOnly = true)
     @Override
-    public User findUserByCar(int carSeries, String carModel) {
+    public User findUserByCar(Long carSeries, String carModel) {
         return userDao.findUserByCar(carSeries, carModel);
     }
